:truck: restructure game-info components under menu-button directory #7
