+++ src/components/frame/with-buttons/menu-button/game-info/MainInfo.tsx
@@ -1,8 +1,9 @@
 import * as styles from './index.css';
 
+import type { infoType } from './types';
+
 import ColoredButton from '@/components/button/ColoredButton';
 import SettingTextButton from '@/components/button/SettingTextButton';
-import type { infoType } from '@/components/frame/with-buttons/game-info/types';
 
 interface mainProps {
   onClose?: () => void;
