+++ src/components/frame/with-buttons/menu-button/game-info/index.tsx
