+++ src/components/frame/with-buttons/menu-button/game-info/types.d.ts
